+++ LillyVKR/LillyVKR.csproj
@@ -63,13 +63,13 @@
   <ItemGroup>
     <Compile Include="Calculation.cs" />
     <Compile Include="Category.cs" />
-    <Compile Include="Form1.cs">
+    <Compile Include="Illness.cs" />
+    <Compile Include="MainForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Form1.Designer.cs">
-      <DependentUpon>Form1.cs</DependentUpon>
+    <Compile Include="MainForm.Designer.cs">
+      <DependentUpon>MainForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="Illness.cs" />
     <Compile Include="myDB.cs" />
     <Compile Include="Patient.cs" />
     <Compile Include="PatientSymptomsForm.cs">
@@ -87,11 +87,8 @@
       <DependentUpon>ResultForm.cs</DependentUpon>
     </Compile>
     <Compile Include="Symptom.cs" />
-    <EmbeddedResource Include="Form1.mzn-IR.resx">
-      <DependentUpon>Form1.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="Form1.resx">
-      <DependentUpon>Form1.cs</DependentUpon>
+    <EmbeddedResource Include="MainForm.resx">
+      <DependentUpon>MainForm.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="PatientSymptomsForm.resx">
       <DependentUpon>PatientSymptomsForm.cs</DependentUpon>
